refactor: move config-related codes into config.rs
